app.pyがapp.routeを受けるよう原点回帰
