+++ main/host/src/host.py
@@ -1,7 +1,6 @@
 print("host importing...")
 
 import os, regex as re
-from flask import render_template, request, jsonify, Blueprint
 import importlib.util
 import main.globals as g
 
@@ -147,23 +146,3 @@ def merge_onload_js():
             f.write("\n")
     clear_onload_js()
 
-
-# ------------------ #
-
-request_module = Blueprint("host", __name__, url_prefix="/")
-
-@request_module.route("/")
-def show_interface():
-    client.generate_html()
-    from main.client.src.client import theme_path
-    return render_template("base.html",theme=theme_path)
-
-@request_module.route("/action", methods=["POST"])
-def action():
-    
-    print(f" ! got POST with arg {request.get_json()}")
-    returns = {}
-    for key, value in request.get_json().items():
-        returns[key] = execute_action(key, value) 
-
-    return jsonify(returns)
